Split main.ts into testable pieces, some other refactoring for singleton pattern
