+++ package.json
@@ -75,7 +75,8 @@
     "testMatch": ["**/src/**/*.test.js"],
     "globals": {
       "__HASH_VERSION__": "5.0.0"
-    }
+    },
+    "setupFiles": ["./src/jestSetup.js"]
   },
   "devDependencies": {
     "@babel/cli": "^7.14.5",
