+++ src/common/config/index.ts
@@ -22,6 +22,7 @@ import {
 
 import {UPDATE_TEAMS, GET_CONFIGURATION, UPDATE_CONFIGURATION, GET_LOCAL_CONFIGURATION} from 'common/communication';
 
+import {configPath} from 'main/constants';
 import * as Validator from 'main/Validator';
 import {getDefaultTeamWithTabsFromTeam} from 'common/tabs/TabView';
 import Utils from 'common/utils/util';
@@ -34,8 +35,6 @@ import RegistryConfig, {REGISTRY_READ_EVENT} from './RegistryConfig';
 /**
  * Handles loading and merging all sources of configuration as well as saving user provided config
  */
-
-const configPath = app.getPath('userData') + '/config.json';
 
 export class Config extends EventEmitter {
     configFilePath: string;
@@ -267,6 +266,12 @@ export class Config extends EventEmitter {
     }
     get helpLink() {
         return this.combinedData?.helpLink;
+    }
+    get minimizeToTray() {
+        return this.combinedData?.minimizeToTray;
+    }
+    get lastActiveTeam() {
+        return this.combinedData?.lastActiveTeam;
     }
 
     // initialization/processing methods
