+++ src/main/app/app.ts
@@ -1,0 +1,145 @@
+// Copyright (c) 2016-present Mattermost, Inc. All Rights Reserved.
+// See LICENSE.txt for license information.
+
+import {app, BrowserWindow, Event, dialog, WebContents, Certificate} from 'electron';
+import log from 'electron-log';
+
+import urlUtils from 'common/utils/url';
+
+import CertificateStore from 'main/certificateStore';
+import {destroyTray} from 'main/tray/tray';
+import WindowManager from 'main/windows/windowManager';
+
+import {getDeeplinkingURL, openDeepLink, resizeScreen} from './utils';
+
+const certificateErrorCallbacks = new Map();
+
+//
+// app event handlers
+//
+
+// activate first app instance, subsequent instances will quit themselves
+export function handleAppSecondInstance(event: Event, argv: string[]) {
+    // Protocol handler for win32
+    // argv: An array of the second instance’s (command line / deep linked) arguments
+    const deeplinkingUrl = getDeeplinkingURL(argv);
+    WindowManager.showMainWindow(deeplinkingUrl);
+}
+
+export function handleAppWindowAllClosed() {
+    // On OS X it is common for applications and their menu bar
+    // to stay active until the user quits explicitly with Cmd + Q
+    if (process.platform !== 'darwin') {
+        app.quit();
+    }
+}
+
+export function handleAppBrowserWindowCreated(event: Event, newWindow: BrowserWindow) {
+    // Screen cannot be required before app is ready
+    resizeScreen(newWindow);
+}
+
+export function handleAppWillFinishLaunching() {
+    // Protocol handler for osx
+    app.on('open-url', (event, url) => {
+        log.info(`Handling deeplinking url: ${url}`);
+        event.preventDefault();
+        const deeplinkingUrl = getDeeplinkingURL([url]);
+        if (deeplinkingUrl) {
+            if (app.isReady() && deeplinkingUrl) {
+                openDeepLink(deeplinkingUrl);
+            } else {
+                app.once('ready', () => openDeepLink(deeplinkingUrl));
+            }
+        }
+    });
+}
+
+export function handleAppBeforeQuit() {
+    // Make sure tray icon gets removed if the user exits via CTRL-Q
+    destroyTray();
+    global.willAppQuit = true;
+}
+
+export function handleAppCertificateError(event: Event, webContents: WebContents, url: string, error: string, certificate: Certificate, callback: (isTrusted: boolean) => void) {
+    const parsedURL = urlUtils.parseURL(url);
+    if (!parsedURL) {
+        return;
+    }
+    const origin = parsedURL.origin;
+    if (CertificateStore.isExplicitlyUntrusted(origin)) {
+        event.preventDefault();
+        log.warn(`Ignoring previously untrusted certificate for ${origin}`);
+        callback(false);
+    } else if (CertificateStore.isTrusted(origin, certificate)) {
+        event.preventDefault();
+        callback(true);
+    } else {
+    // update the callback
+        const errorID = `${origin}:${error}`;
+
+        // if we are already showing that error, don't add more dialogs
+        if (certificateErrorCallbacks.has(errorID)) {
+            log.warn(`Ignoring already shown dialog for ${errorID}`);
+            certificateErrorCallbacks.set(errorID, callback);
+            return;
+        }
+        const extraDetail = CertificateStore.isExisting(origin) ? 'Certificate is different from previous one.\n\n' : '';
+        const detail = `${extraDetail}origin: ${origin}\nError: ${error}`;
+
+        certificateErrorCallbacks.set(errorID, callback);
+
+        // TODO: should we move this to window manager or provide a handler for dialogs?
+        const mainWindow = WindowManager.getMainWindow();
+        if (!mainWindow) {
+            return;
+        }
+        dialog.showMessageBox(mainWindow, {
+            title: 'Certificate Error',
+            message: 'There is a configuration issue with this Mattermost server, or someone is trying to intercept your connection. You also may need to sign into the Wi-Fi you are connected to using your web browser.',
+            type: 'error',
+            detail,
+            buttons: ['More Details', 'Cancel Connection'],
+            cancelId: 1,
+        }).then(
+            ({response}) => {
+                if (response === 0) {
+                    return dialog.showMessageBox(mainWindow, {
+                        title: 'Certificate Not Trusted',
+                        message: `Certificate from "${certificate.issuerName}" is not trusted.`,
+                        detail: extraDetail,
+                        type: 'error',
+                        buttons: ['Trust Insecure Certificate', 'Cancel Connection'],
+                        cancelId: 1,
+                        checkboxChecked: false,
+                        checkboxLabel: "Don't ask again",
+                    });
+                }
+                return {response, checkboxChecked: false};
+            }).then(
+            ({response: responseTwo, checkboxChecked}) => {
+                if (responseTwo === 0) {
+                    CertificateStore.add(origin, certificate);
+                    CertificateStore.save();
+                    certificateErrorCallbacks.get(errorID)(true);
+                    certificateErrorCallbacks.delete(errorID);
+                    webContents.loadURL(url);
+                } else {
+                    if (checkboxChecked) {
+                        CertificateStore.add(origin, certificate, true);
+                        CertificateStore.save();
+                    }
+                    certificateErrorCallbacks.get(errorID)(false);
+                    certificateErrorCallbacks.delete(errorID);
+                }
+            }).catch(
+            (dialogError) => {
+                log.error(`There was an error with the Certificate Error dialog: ${dialogError}`);
+                certificateErrorCallbacks.delete(errorID);
+            });
+    }
+}
+
+export function handleAppGPUProcessCrashed(event: Event, killed: boolean) {
+    log.error(`The GPU process has crashed (killed = ${killed})`);
+}
